--- api/src/main/java/io/eagle/domain/vacation/dto/LatestCahootsDto.java
@@ -1,19 +1,0 @@
-package io.eagle.domain.vacation.dto;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-public class LatestCahootsDto {
-    private Long id;
-    private String title;
-
-    private LocalDate stockStart;
-    private Integer expectedRateOfReturn;
-
-    private List<String> images;
-}
