+++ api/src/main/java/io/eagle/domain/vacation/dto/response/BreifCahootsDto.java
@@ -1,0 +1,33 @@
+package io.eagle.domain.vacation.dto.response;
+
+import io.eagle.entity.type.VacationStatusType;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+import java.util.List;
+
+@Data
+@NoArgsConstructor
+public class BreifCahootsDto {
+    private Long id;
+    private String title;
+    private String location;
+    private VacationStatusType status;
+
+    private LocalDate stockStart;
+
+    private LocalDate stockEnd;
+
+    private Long stockPrice;
+
+    private Integer stockNum;
+
+    private Integer competitionRate;
+
+    // TODO : hashtag
+
+    // TODO : picture url
+    private List<String> images;
+
+}
