+++ api/src/main/java/io/eagle/domain/vacation/dto/response/LatestCahootsDto.java
@@ -1,6 +1,5 @@
-package io.eagle.domain.vacation.dto;
+package io.eagle.domain.vacation.dto.response;
 
-import io.eagle.entity.type.VacationStatusType;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -9,15 +8,12 @@ import java.util.List;
 
 @Data
 @NoArgsConstructor
-public class BreifV2CahootsDto {
+public class LatestCahootsDto {
     private Long id;
     private String title;
-    private VacationStatusType status;
 
     private LocalDate stockStart;
-
-    private LocalDate stockEnd;
+    private Integer expectedRateOfReturn;
 
     private List<String> images;
-
 }
