+++ api/src/main/java/io/eagle/domain/vacation/repository/VacationCustom.java
@@ -1,6 +1,10 @@
 package io.eagle.domain.vacation.repository;
 
 import io.eagle.domain.vacation.dto.*;
+import io.eagle.domain.vacation.dto.response.BreifCahootsDto;
+import io.eagle.domain.vacation.dto.response.BreifV2CahootsDto;
+import io.eagle.domain.vacation.dto.response.DetailCahootsDto;
+import io.eagle.domain.vacation.dto.response.LatestCahootsDto;
 import io.eagle.entity.Vacation;
 import org.springframework.data.domain.Pageable;
