+++ api/src/main/java/io/eagle/domain/vacation/repository/VacationCustomImpl.java
@@ -5,6 +5,10 @@ import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.JPQLQueryFactory;
 import io.eagle.domain.vacation.dto.*;
+import io.eagle.domain.vacation.dto.response.BreifCahootsDto;
+import io.eagle.domain.vacation.dto.response.BreifV2CahootsDto;
+import io.eagle.domain.vacation.dto.response.DetailCahootsDto;
+import io.eagle.domain.vacation.dto.response.LatestCahootsDto;
 import io.eagle.entity.Vacation;
 import io.eagle.entity.type.VacationStatusType;
 import lombok.RequiredArgsConstructor;
