+++ api/src/main/java/io/eagle/domain/vacation/service/CahootsService.java
@@ -1,6 +1,8 @@
 package io.eagle.domain.vacation.service;
 
 import io.eagle.domain.picture.S3;
+import io.eagle.domain.vacation.dto.request.CreateCahootsDto;
+import io.eagle.domain.vacation.dto.response.*;
 import io.eagle.entity.Picture;
 import io.eagle.domain.picture.repository.PictureRepository;
 import io.eagle.domain.vacation.dto.*;
